Made a loop with a check at the end
COnverted the model into a loop that will run for each species and that produces a data frame that you can check at the end. WIll probably want to save that data frame but for now it is just to check convergence. Also had to add more iterations to be sure of convergence