Fixed the math to what we want detailed in the description
Now we have a genus level slope that is multiplied by the latitude. That is added to the species effect. The model works and converges much better than I expected. I'm a bit worried about the biology since the Species effects are around 1000 and the slope and latitude effect brings it down to normal levels. Unsure if that's something to worry about since we get to biological logical answers.